pyproject.toml
[tool.poetry]
name = "ser-gawain"
version = "0.1.0"
description = ""
authors = ["Nick <9609041+0niSec@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "^2.4.0"
python-dotenv = "^1.0.1"
tabulate = "^0.9.0"
asynctest = "^0.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
